provide an example of a working RMSE and PRFS
