+++ src/examples/timer.py
@@ -1,0 +1,16 @@
+import time
+
+class Timer(object):
+    def __init__(self, verbose=False):
+        self.verbose = verbose
+
+    def __enter__(self):
+        self.start = time.time()
+        return self
+
+    def __exit__(self, *args):
+        self.end = time.time()
+        self.secs = self.end - self.start
+        self.msecs = self.secs * 1000
+        if self.verbose:
+            print "elapsed time: %f ms" % self.msecs
